src/api/layers/common/requirements.txt
alembic==1.13.1
aws-lambda-powertools[all]
bcrypt==4.1.2
openai==1.6.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic[email]
PyJWT==2.8.0
sqlalchemy==2.0.25